simplify build process upon review
