+++ boilerplates/webpack-with-jsx/.gitignore
@@ -2,4 +2,3 @@ bower_components
 build
 coverage
 node_modules
-stat.html
