+++ boilerplates/webpack-with-jsx/package.json
@@ -1,6 +1,5 @@
 {
 	"name": "metal-modal",
-	"publicName": "Modal",
 	"version": "0.0.0",
 	"description": "My awesome Metal.js project",
 	"license": "BSD",
@@ -15,8 +14,7 @@
 		"test"
 	],
 	"scripts": {
-		"start": "node scripts/start.js",
-		"build": "node scripts/build.js"
+		"build": "webpack"
 	},
 	"keywords": [
 		"metal"
@@ -33,17 +31,6 @@
 		"babel-preset-metal": "^4.1.0",
 		"babel-preset-metal-jsx": "^0.1.1",
 		"babel-preset-react": "^6.24.1",
-		"css-loader": "^0.28.4",
-		"detect-port": "^1.2.1",
-		"fs-extra": "^4.0.1",
-		"inquirer": "^3.2.2",
-		"node-sass": "^4.5.3",
-		"react-dev-utils": "^3.1.1",
-		"react-hot-loader": "^1.3.1",
-		"sass-loader": "^6.0.6",
-		"style-loader": "^0.18.2",
-		"webpack": "^3.5.4",
-		"webpack-dev-server": "^2.7.1",
-		"webpack-visualizer-plugin": "^0.1.11"
+		"webpack": "^3.5.4"
 	}
 }
